restart worker during tests depending on resident memory size
(cherry picked from commit 17e2e1c42855578ab6c558248365c49bc4d2f55f)
ref #13577
