+++ test/examples.jl
@@ -51,11 +51,10 @@ end
 dc_path = joinpath(dir, "dictchannel.jl")
 include(dc_path)
 
-w_set=filter!(x->x != myid(), workers())
-pid = length(w_set) > 0 ? w_set[1] : myid()
-
-remotecall_fetch(pid, f->(include(f); nothing), dc_path)
-dc=RemoteRef(()->DictChannel(), pid)
+# Run the remote on pid 1, since runtests may terminate workers
+# at any time depending on memory usage
+remotecall_fetch(1, f->(include(f); nothing), dc_path)
+dc=RemoteRef(()->DictChannel(), 1)
 @test typeof(dc) == RemoteRef{DictChannel}
 
 @test isready(dc) == false
