+++ test/testdefs.jl
@@ -5,18 +5,9 @@ using Base.Test
 function runtests(name)
     @printf("     \033[1m*\033[0m \033[31m%-21s\033[0m", name)
     tt = @elapsed include("$name.jl")
-    @printf(" in %6.2f seconds\n", tt)
-    nothing
-end
-
-function propagate_errors(a,b)
-    if isa(a,Exception)
-        rethrow(a)
-    end
-    if isa(b,Exception)
-        rethrow(b)
-    end
-    nothing
+    rss = Sys.maxrss()
+    @printf(" in %6.2f seconds, maxrss %7.2f MB\n", tt, rss / 2^20)
+    rss
 end
 
 # looking in . messes things up badly
